feat(env-utils): move os name logic from suite package
