+++ packages/suite/src/utils/suite/__tests__/messageSystem.test.ts
@@ -1,6 +1,7 @@
 import * as messageSystem from '../messageSystem';
 import * as fixtures from '../__fixtures__/messageSystem';
 import * as env from '@suite-utils/env';
+import * as envUtils from '@trezor/env-utils/src/envUtils';
 
 describe('Message system utils', () => {
     describe('createVersionRange', () => {
@@ -110,7 +111,7 @@ describe('Message system utils', () => {
             it(f.description, () => {
                 jest.spyOn(Date, 'now').mockImplementation(() => new Date(f.currentDate).getTime());
                 // @ts-expect-error
-                jest.spyOn(env, 'getOsName').mockImplementation(() => f.osName);
+                jest.spyOn(envUtils, 'getOsName').mockImplementation(() => f.osName);
                 userAgentGetter.mockReturnValue(f.userAgent);
                 // @ts-expect-error
                 jest.spyOn(env, 'getEnvironment').mockImplementation(() => f.environment);
